Move TxnWriter to posting pkg, so it can discard earlier versions based on Posting bits.
