+++ posting/mvcc.go
@@ -86,7 +86,7 @@ func (t *Txn) Fill(ctx *api.TxnContext) {
 // generate state. The state should only be generated by rollup, which in turn should look at the
 // last Snapshot Ts, to determine how much of the PL to rollup. We only want to roll up the deltas,
 // with commit ts <= snapshot ts, and not above.
-func (tx *Txn) CommitToDisk(writer *x.TxnWriter, commitTs uint64) error {
+func (tx *Txn) CommitToDisk(writer *TxnWriter, commitTs uint64) error {
 	if commitTs == 0 {
 		return nil
 	}
